requirements.txt
asgiref==3.8.1
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
django-libsass==0.9
django-simple-history==3.7.0
django_compressor==4.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
libsass==0.23.0
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
